requirements.txt
Flask==3.0.0
psycopg2-binary
anthropic==0.39.0
python-dotenv==1.0.0
